single view file
